Rough Header and Footers
Here is the rough header and footer docs. I will clean it up and try to
get it as true to the wireframe as I can.

I'm using a skinny browser window to approximate the size of an xs
mobile screen, so the proportions might not be exactly right. Luckily it
wont be too difficult to go through and make edits because there isn't a
whole lot of code yet.

I am approaching this kind of like a wireframe, keeping everything in
grayscale, etc. because we haven't really made any comps to determine
design style yet.

Let me know if this is coming along well.